+++ logs.txt
@@ -1,0 +1,5 @@
+UncaughtException error:  Error: UncaughtException!!!
+    at file:///D:/WorkJS/RS_School_Course_NodeJS2021/build/app.js:49:7
+    at ModuleJob.run (internal/modules/esm/module_job.js:152:23)
+    at async Loader.import (internal/modules/esm/loader.js:166:24)
+    at async Object.loadESM (internal/process/esm_loader.js:68:5)
