--- src/errors/logger.ts
@@ -1,7 +1,0 @@
-export default (text: string, err?: Error) => {
-    if (err) {
-        console.log(`${text} ${err.stack}`);
-    } else {
-        console.log(text);
-    }
-};
